Fix duplicated edges problem in Java Topology.getAllEdges()
If a topology is undirected The Topology.getAllEdges() was
returning each edge twice, first in the form (u,v) and then in
the form (v, u). This commit prevents this from happening again.

It also fixes Topology.clone() method which was deep-copying nodes
and edges.

This commit also includes new test cases validating the features
added.