Improves TM calculation with sparse origins (close #5)
The identification of OD pairs used to be carried out by running the
Dijkstra algorithms between all pairs of nodes, even if only a small
subset of nodes were selected as origins. Now the Dijkstra algorithm
is calculated only on the selected origins.

Note that if TM generation functions are invoked specifying (not None)
origin_nodes and destination_nodes arguments, there is no control on
the actual connectivity of origins and destination. The user is
responsible for ensuring this. This behavior may be changed in future
but needs to be thought more carefully because the performance
overhead for this operation may be high in very large topologies.

This commit also fixes some problems occurring in the NFUR calculation
due to the transformation of a topology between directed and
undirected which sometimes originated a KeyError when trying to access
specific edges.