Add ns-3 adapter
This commit adds the ns-3 adapter.

At this stage, the ns-3 adapter is capable of correctly parsing both
topology and event schedule XML files and run an ns-3 simulation
based on them.

This commit adds source code, documentation, examples and build
scripts. However, the documentation is not very comprehensive.
It will be improved by a future commit.