Fix infinite loop in GLP topology generator
This commit fixes a defect that occasionally originated an infinite
loop while building a GLP topology.

If while bulding the topology, the generator found itself in the
condition of being unable to add m links (because near full mesh),
the generator kept trying adding links anyway (unsuccessfully).

Now, the generator, if the same situation occurs, adds nodes util it
is again possible to add m links or until the target topology size is
reached.