Change C++ and ns-3 Quantity::toString() signature
This commit changes the signature of Quantity::toString() function,
present in both C++ and ns-3 components, so that it now returns by
default the value and the unit concatenated without spaces.

This fixes a defect in the ns-3 adapter which was not able to
recognise capacities and delays with a space between value and unit.