Improve set_capacities_random_power_law function
This commit improves the function assigning link capacities according
to a power law distribution. Before this commit, this function
assigned capacity so that the product of capacity value and
probability that a link has that capacity is constant, i.e. the
skewness parameter of the power-law relationship was fixed to 1.

After this commit, capacity value and probability have a power-law
relationship with tunable skewness.