Replace calls to type with isinstance
This commit replaces in all Python code calls to the 'type' function
with calls to 'isinstance' when used in comparison. For example,
'if type(x) is int' becomes 'if isinstance(x, int)' with this commit.
This changes has been made because the use of type in such situations
is deprecated.