Move ns-2 adapter inside the core library
This commit moves the code for exporting FNSS topologies to ns-2 from
a dedicated library to the core library.

This change has been done because, since the code is entirely written
in Python, it is easier for the user to have it integrated in the
core library as opposed to having to deal with an additional
component.

This change also required to change template engine from Templite++
to Mako because of licensing reasons. Templite++ is released
under the GPLv2 license which does not allow it being used in a
BSD-licensed project (like the FNSS core library). Differently,
Mako is released under the MIT license which is fine.