Automate document generation in core library
This commit adds new scripts and modifies existing code to automate
almost entirely the generation of documentation for the core library.

Before this commit, changes in the code required quite few manual
changes to the source code of the documentation. For example, changes
in the examples needed to be encoded manually in the documentation
as well as changes in API definition.

This commit automates all this.

There is only one bit that has not been automated at this stage:
the list of classes and functions included in the library. This
has been done to leave freedom to the developer to give freedom
to decide how to organize the documentation of various functions
and classes. This may change in future though.