Update topology file of ns-3 example
This commit updates and adds comments to the topology XML file of
the ns-3 library example.

This commit however does not fix entirely a bug that makes the ns-3
example in versions starting from 3.18 crash due to a failure in
handling IPv4 addresses correctly by ns-3. I attempted to fix the
issue but did not address it completely yet. It will be addressed
by a later commit.