Support lists of only one capacity value in capacity setters
If a user tried to assign capacities proportionally to a centrality
metric using a capacities list of only one value, would have
raised an IndexError, as the _set_capacities_proportionally function
assumed that the capacities list had at least two values.

Now, _set_capacities_proportionally checks if the capacities list
contains only one element, and if so, assigns to all links of the
topology the same capacity value, which is contained in the
capacities list.

This commit fixes #16.